backend/requirements.txt
django
pymysql
djangorestframework
xlwt
pillow
pyjwt==1.7.1